refactor: move signature utils to separate file in devkit
